Fix warning in test code

+++ src/base/SbByteBuffer.cpp
@@ -62,9 +62,9 @@ BOOST_AUTO_TEST_CASE(pushOnEmpty)
                       "Size of buffers differ"
                       );
 
-  for (int i=0; i < b.size(); ++i) {
+  for (size_t i=0; i < b.size(); ++i) {
     if(a[i]!=b[i]) {
-      printf("Mjau %d: %c != %c \n",i,a[i],b[i]);
+      printf("Mjau %lu: %c != %c \n",(unsigned long)i,a[i],b[i]);
     }
   }
